movie/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
alembic
cryptography
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-jaeger
dapr
pyodbc
graphene 
graphene-sqlalchemy
graphql-core
strawberry-graphql[fastapi]